+++ Localization/enUS.lua
@@ -139,6 +139,10 @@ L.Settings.ProgressBarColorTooltip =
 
 L.Settings.MirrorLayoutLabel = "Mirror Layout"
 
+L.Settings.AnnounceUntargetedSpellsLabel = "Announce Untargeted Spells"
+L.Settings.AnnounceUntargetedSpellsTooltip =
+	"Plays Text-To-Speech for any spell that doesn't target anyone, such as AoE and some frontals."
+
 L.Settings.HideUntargetedSpellsLabel = "Hide Untargeted Spells"
 
 L.Settings.HideTargetedSpellsLabel = "Hide Targeted Spells"